requirements.txt

fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
aiosmtplib==2.0.1
python-magic==0.4.27